+++ src/test/java/io/hhplus/tdd/point/PointServiceIntegrationTest.java
@@ -1,5 +1,7 @@
-package io.hhplus.tdd.database;
+package io.hhplus.tdd.point;
 
+import io.hhplus.tdd.database.PointHistoryTable;
+import io.hhplus.tdd.database.UserPointTable;
 import io.hhplus.tdd.point.PointPolicy;
 import io.hhplus.tdd.point.PointService;
 import io.hhplus.tdd.point.TransactionType;
